chore(derive): rebase + move `alloy` module
