+++ crates/derive/src/types/alloy/transaction/eip4844.rs
@@ -1,34 +1,46 @@
 use crate::types::{
-    eips::eip2930::AccessList,
-    network::{Signed, Transaction, TxKind},
-    transaction::TxType,
+    eip2930::AccessList, eip4844::DATA_GAS_PER_BLOB, Signed, Transaction, TxKind, TxType,
 };
 use alloc::vec::Vec;
-use alloy_primitives::{keccak256, Bytes, ChainId, Signature, U256};
+use alloy_primitives::{keccak256, Bytes, ChainId, Signature, B256, U256};
 use alloy_rlp::{length_of_length, BufMut, Decodable, Encodable, Header};
 use core::mem;
 
-/// Transaction with an [`AccessList`] ([EIP-2930](https://eips.ethereum.org/EIPS/eip-2930)).
+/// [EIP-4844 Blob Transaction](https://eips.ethereum.org/EIPS/eip-4844#blob-transaction)
+///
+/// A transaction with blob hashes and max blob fee
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
-pub struct TxEip2930 {
+pub struct TxEip4844 {
     /// Added as EIP-pub 155: Simple replay attack protection
     pub chain_id: ChainId,
     /// A scalar value equal to the number of transactions sent by the sender; formally Tn.
     pub nonce: u64,
-    /// A scalar value equal to the number of
-    /// Wei to be paid per unit of gas for all computation
-    /// costs incurred as a result of the execution of this transaction; formally Tp.
-    ///
-    /// As ethereum circulation is around 120mil eth as of 2022 that is around
-    /// 120000000000000000000000000 wei we are safe to use u128 as its max number is:
-    /// 340282366920938463463374607431768211455
-    pub gas_price: u128,
     /// A scalar value equal to the maximum
     /// amount of gas that should be used in executing
     /// this transaction. This is paid up-front, before any
     /// computation is done and may not be increased
     /// later; formally Tg.
     pub gas_limit: u64,
+    /// A scalar value equal to the maximum
+    /// amount of gas that should be used in executing
+    /// this transaction. This is paid up-front, before any
+    /// computation is done and may not be increased
+    /// later; formally Tg.
+    ///
+    /// As ethereum circulation is around 120mil eth as of 2022 that is around
+    /// 120000000000000000000000000 wei we are safe to use u128 as its max number is:
+    /// 340282366920938463463374607431768211455
+    ///
+    /// This is also known as `GasFeeCap`
+    pub max_fee_per_gas: u128,
+    /// Max Priority fee that transaction is paying
+    ///
+    /// As ethereum circulation is around 120mil eth as of 2022 that is around
+    /// 120000000000000000000000000 wei we are safe to use u128 as its max number is:
+    /// 340282366920938463463374607431768211455
+    ///
+    /// This is also known as `GasTipCap`
+    pub max_priority_fee_per_gas: u128,
     /// The 160-bit address of the message call’s recipient or, for a contract creation
     /// transaction, ∅, used here to denote the only member of B0 ; formally Tt.
     pub to: TxKind,
@@ -43,6 +55,15 @@ pub struct TxEip2930 {
     /// A gas cost is charged, though at a discount relative to the cost of
     /// accessing outside the list.
     pub access_list: AccessList,
+
+    /// It contains a vector of fixed size hash(32 bytes)
+    pub blob_versioned_hashes: Vec<B256>,
+
+    /// Max fee per data gas
+    ///
+    /// aka BlobFeeCap or blobGasFeeCap
+    pub max_fee_per_blob_gas: u128,
+
     /// Input has two uses depending if transaction is Create or Call (if `to` field is None or
     /// Some). pub init: An unlimited size byte array specifying the
     /// EVM-code for the account initialisation procedure CREATE,
@@ -51,43 +72,61 @@ pub struct TxEip2930 {
     pub input: Bytes,
 }
 
-impl TxEip2930 {
-    /// Calculates a heuristic for the in-memory size of the [TxEip2930] transaction.
-    #[inline]
-    pub fn size(&self) -> usize {
-        mem::size_of::<ChainId>() + // chain_id
-        mem::size_of::<u64>() + // nonce
-        mem::size_of::<u128>() + // gas_price
-        mem::size_of::<u64>() + // gas_limit
-        self.to.size() + // to
-        mem::size_of::<U256>() + // value
-        self.access_list.size() + // access_list
-        self.input.len() // input
+impl TxEip4844 {
+    /// Returns the effective gas price for the given `base_fee`.
+    pub const fn effective_gas_price(&self, base_fee: Option<u64>) -> u128 {
+        match base_fee {
+            None => self.max_fee_per_gas,
+            Some(base_fee) => {
+                // if the tip is greater than the max priority fee per gas, set it to the max
+                // priority fee per gas + base fee
+                let tip = self.max_fee_per_gas.saturating_sub(base_fee as u128);
+                if tip > self.max_priority_fee_per_gas {
+                    self.max_priority_fee_per_gas + base_fee as u128
+                } else {
+                    // otherwise return the max fee per gas
+                    self.max_fee_per_gas
+                }
+            }
+        }
     }
 
-    /// Decodes the inner [TxEip2930] fields from RLP bytes.
+    /// Returns the total gas for all blobs in this transaction.
+    #[inline]
+    pub fn blob_gas(&self) -> u64 {
+        // SAFETY: we don't expect u64::MAX / DATA_GAS_PER_BLOB hashes in a single transaction
+        self.blob_versioned_hashes.len() as u64 * DATA_GAS_PER_BLOB
+    }
+
+    /// Decodes the inner [TxEip4844] fields from RLP bytes.
     ///
     /// NOTE: This assumes a RLP header has already been decoded, and _just_ decodes the following
     /// RLP fields in the following order:
     ///
     /// - `chain_id`
     /// - `nonce`
-    /// - `gas_price`
+    /// - `max_priority_fee_per_gas`
+    /// - `max_fee_per_gas`
     /// - `gas_limit`
     /// - `to`
     /// - `value`
     /// - `data` (`input`)
     /// - `access_list`
-    pub(crate) fn decode_inner(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
+    /// - `max_fee_per_blob_gas`
+    /// - `blob_versioned_hashes`
+    pub fn decode_inner(buf: &mut &[u8]) -> alloy_rlp::Result<Self> {
         Ok(Self {
             chain_id: Decodable::decode(buf)?,
             nonce: Decodable::decode(buf)?,
-            gas_price: Decodable::decode(buf)?,
+            max_priority_fee_per_gas: Decodable::decode(buf)?,
+            max_fee_per_gas: Decodable::decode(buf)?,
             gas_limit: Decodable::decode(buf)?,
             to: Decodable::decode(buf)?,
             value: Decodable::decode(buf)?,
             input: Decodable::decode(buf)?,
             access_list: Decodable::decode(buf)?,
+            max_fee_per_blob_gas: Decodable::decode(buf)?,
+            blob_versioned_hashes: Decodable::decode(buf)?,
         })
     }
 
@@ -96,12 +135,15 @@ impl TxEip2930 {
         let mut len = 0;
         len += self.chain_id.length();
         len += self.nonce.length();
-        len += self.gas_price.length();
         len += self.gas_limit.length();
+        len += self.max_fee_per_gas.length();
+        len += self.max_priority_fee_per_gas.length();
         len += self.to.length();
         len += self.value.length();
-        len += self.input.0.length();
         len += self.access_list.length();
+        len += self.blob_versioned_hashes.length();
+        len += self.max_fee_per_blob_gas.length();
+        len += self.input.0.length();
         len
     }
 
@@ -109,32 +151,57 @@ impl TxEip2930 {
     pub(crate) fn encode_fields(&self, out: &mut dyn BufMut) {
         self.chain_id.encode(out);
         self.nonce.encode(out);
-        self.gas_price.encode(out);
+        self.max_priority_fee_per_gas.encode(out);
+        self.max_fee_per_gas.encode(out);
         self.gas_limit.encode(out);
         self.to.encode(out);
         self.value.encode(out);
         self.input.0.encode(out);
         self.access_list.encode(out);
+        self.max_fee_per_blob_gas.encode(out);
+        self.blob_versioned_hashes.encode(out);
+    }
+
+    /// Calculates a heuristic for the in-memory size of the [TxEip4844] transaction.
+    #[inline]
+    pub fn size(&self) -> usize {
+        mem::size_of::<ChainId>() + // chain_id
+        mem::size_of::<u64>() + // nonce
+        mem::size_of::<u64>() + // gas_limit
+        mem::size_of::<u128>() + // max_fee_per_gas
+        mem::size_of::<u128>() + // max_priority_fee_per_gas
+        self.to.size() + // to
+        mem::size_of::<U256>() + // value
+        self.access_list.size() + // access_list
+        self.input.len() +  // input
+        self.blob_versioned_hashes.capacity() * mem::size_of::<B256>() + // blob hashes size
+        mem::size_of::<u128>() // max_fee_per_data_gas
     }
 
     /// Inner encoding function that is used for both rlp [`Encodable`] trait and for calculating
     /// hash that for eip2718 does not require rlp header
-    pub(crate) fn encode_with_signature(&self, signature: &Signature, out: &mut dyn BufMut) {
+    pub(crate) fn encode_with_signature(
+        &self,
+        signature: &Signature,
+        out: &mut dyn BufMut,
+        with_header: bool,
+    ) {
         let payload_length = self.fields_len() + signature.rlp_vrs_len();
+        if with_header {
+            Header {
+                list: false,
+                payload_length: 1 + length_of_length(payload_length) + payload_length,
+            }
+            .encode(out);
+        }
+        out.put_u8(self.tx_type() as u8);
         let header = Header {
             list: true,
             payload_length,
         };
         header.encode(out);
         self.encode_fields(out);
-        signature.write_rlp_vrs(out);
-    }
-
-    /// Output the length of the RLP signed transaction encoding, _without_ a RLP string header.
-    pub fn payload_len_with_signature_without_header(&self, signature: &Signature) -> usize {
-        let payload_length = self.fields_len() + signature.rlp_vrs_len();
-        // 'transaction type byte length' + 'header length' + 'payload length'
-        1 + length_of_length(payload_length) + payload_length
+        signature.encode(out);
     }
 
     /// Output the length of the RLP signed transaction encoding. This encodes with a RLP header.
@@ -143,46 +210,26 @@ impl TxEip2930 {
         length_of_length(len) + len
     }
 
-    /// Get transaction type.
+    /// Output the length of the RLP signed transaction encoding, _without_ a RLP header.
+    pub fn payload_len_with_signature_without_header(&self, signature: &Signature) -> usize {
+        let payload_length = self.fields_len() + signature.rlp_vrs_len();
+        // 'transaction type byte length' + 'header length' + 'payload length'
+        1 + length_of_length(payload_length) + payload_length
+    }
+
+    /// Get transaction type
     pub const fn tx_type(&self) -> TxType {
-        TxType::Eip2930
-    }
-}
-
-impl Encodable for TxEip2930 {
-    fn encode(&self, out: &mut dyn BufMut) {
-        Header {
-            list: true,
-            payload_length: self.fields_len(),
-        }
-        .encode(out);
-        self.encode_fields(out);
+        TxType::Eip4844
     }
 
-    fn length(&self) -> usize {
-        let payload_length = self.fields_len();
-        length_of_length(payload_length) + payload_length
-    }
-}
-
-impl Decodable for TxEip2930 {
-    fn decode(data: &mut &[u8]) -> alloy_rlp::Result<Self> {
-        let header = Header::decode(data)?;
-        let remaining_len = data.len();
-
-        if header.payload_length > remaining_len {
-            return Err(alloy_rlp::Error::InputTooShort);
-        }
-
-        Self::decode_inner(data)
-    }
-}
-
-impl Transaction for TxEip2930 {
-    type Signature = Signature;
-    // type Receipt = ReceiptWithBloom;
-
-    fn encode_for_signing(&self, out: &mut dyn BufMut) {
+    /// Encodes the legacy transaction in RLP for signing.
+    ///
+    /// This encodes the transaction as:
+    /// `tx_type || rlp(chain_id, nonce, max_priority_fee_per_gas, max_fee_per_gas, gas_limit, to,
+    /// value, input, access_list, max_fee_per_blob_gas, blob_versioned_hashes)`
+    ///
+    /// Note that there is no rlp header before the transaction type byte.
+    pub fn encode_for_signing(&self, out: &mut dyn BufMut) {
         out.put_u8(self.tx_type() as u8);
         Header {
             list: true,
@@ -190,6 +237,21 @@ impl Transaction for TxEip2930 {
         }
         .encode(out);
         self.encode_fields(out);
+    }
+
+    /// Outputs the length of the signature RLP encoding for the transaction.
+    pub fn payload_len_for_signature(&self) -> usize {
+        let payload_length = self.fields_len();
+        // 'transaction type byte length' + 'header length' + 'payload length'
+        1 + length_of_length(payload_length) + payload_length
+    }
+}
+
+impl Transaction for TxEip4844 {
+    type Signature = Signature;
+
+    fn chain_id(&self) -> Option<ChainId> {
+        Some(self.chain_id)
     }
 
     fn payload_len_for_signature(&self) -> usize {
@@ -201,18 +263,14 @@ impl Transaction for TxEip2930 {
     fn into_signed(self, signature: Signature) -> Signed<Self> {
         let payload_length = 1 + self.fields_len() + signature.rlp_vrs_len();
         let mut buf = Vec::with_capacity(payload_length);
-        buf.put_u8(TxType::Eip2930 as u8);
+        buf.put_u8(TxType::Eip1559 as u8);
         self.encode_signed(&signature, &mut buf);
         let hash = keccak256(&buf);
 
         // Drop any v chain id value to ensure the signature format is correct at the time of
-        // combination for an EIP-2930 transaction. V should indicate the y-parity of the
+        // combination for an EIP-4844 transaction. V should indicate the y-parity of the
         // signature.
         Signed::new_unchecked(self, signature.with_parity_bool(), hash)
-    }
-
-    fn encode_signed(&self, signature: &Signature, out: &mut dyn BufMut) {
-        self.encode_with_signature(signature, out)
     }
 
     fn decode_signed(buf: &mut &[u8]) -> alloy_rlp::Result<Signed<Self>> {
@@ -225,6 +283,14 @@ impl Transaction for TxEip2930 {
         let signature = Signature::decode_rlp_vrs(buf)?;
 
         Ok(tx.into_signed(signature))
+    }
+
+    fn encode_for_signing(&self, out: &mut dyn alloy_rlp::BufMut) {
+        self.encode_for_signing(out);
+    }
+
+    fn encode_signed(&self, signature: &Signature, out: &mut dyn BufMut) {
+        TxEip4844::encode_with_signature(self, signature, out, true);
     }
 
     fn input(&self) -> &[u8] {
@@ -255,10 +321,6 @@ impl Transaction for TxEip2930 {
         self.value = value;
     }
 
-    fn chain_id(&self) -> Option<ChainId> {
-        Some(self.chain_id)
-    }
-
     fn set_chain_id(&mut self, chain_id: ChainId) {
         self.chain_id = chain_id;
     }
@@ -280,12 +342,10 @@ impl Transaction for TxEip2930 {
     }
 
     fn gas_price(&self) -> Option<U256> {
-        Some(U256::from(self.gas_price))
+        None
     }
 
     fn set_gas_price(&mut self, price: U256) {
-        if let Ok(price) = price.try_into() {
-            self.gas_price = price;
-        }
+        let _ = price;
     }
 }
